% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_age}
\alias{validate_age}
\title{Validate age column. Acceptable ages are: 0,1,2,4,5,6
--flag any records with blank age}
\usage{
validate_age(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with age issues
}
\description{
Validate age column. Acceptable ages are: 0,1,2,4,5,6
--flag any records with blank age
}
