% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_age_ffmlt}
\alias{validate_age_ffmlt}
\title{Validate age-ffmolt combinations. Blanks are okay, and can match with any
age. Refer to table on rules page}
\usage{
validate_age_ffmlt(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with age/ffmolt issues
}
\description{
Validate age-ffmolt combinations. Blanks are okay, and can match with any
age. Refer to table on rules page
}
