% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_age_ffwear}
\alias{validate_age_ffwear}
\title{Validate age-ffwear combinations.}
\usage{
validate_age_ffwear(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with age/ffwear issues
}
\description{
0 or 1 FF wear is highly suspicious for age 5 and 6. Flag all these records
Sometimes 0 FF wear is normal if paired with S FF molt, but then
micro-ageing is suspect, so we should flag the record either way, maybe with
a message FF wear and age combination unlikely. Check this record
}
\details{
2+ FF wear is suspicious for age 4--add message age and FF wear combination
unlikely

4+ is suspicious for age 2--add unlikely message
}
