% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_age_ha}
\alias{validate_age_ha}
\title{Validate age-how aged combinations}
\usage{
validate_age_ha(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with age/how aged issues
}
\description{
Validate age-how aged combinations
}
