% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_age_hs}
\alias{validate_age_hs}
\title{Validate age-how sexed combinations}
\usage{
validate_age_hs(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with age/how sexed issues
}
\description{
Validate age-how sexed combinations
}
