% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_age_skull}
\alias{validate_age_skull}
\title{Validate age and skull combinations. Allowable values for skull 0-6, 8,9,
blank. Flag all values in the skull column that don't match these}
\usage{
validate_age_skull(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with age/skull issues
}
\description{
Validate age and skull combinations. Allowable values for skull 0-6, 8,9,
blank. Flag all values in the skull column that don't match these
}
