% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_all_list}
\alias{validate_all_list}
\title{Validate all columns, then store issues as a list of data frames}
\usage{
validate_all_list(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
list of issue data frames
}
\description{
Validate all columns, then store issues as a list of data frames
}
