% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bandcode}
\alias{validate_bandcode}
\title{Validate band code. Make sure there are no
blanks. Make sure the only values used are 1,R,4,5,8,N,U.}
\usage{
validate_bandcode(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with band code issues
}
\description{
Validate band code. Make sure there are no
blanks. Make sure the only values used are 1,R,4,5,8,N,U.
}
