% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bandcode_species}
\alias{validate_bandcode_species}
\title{Validate band code-species combinations. Make sure 4 and 8
are only used for species codes BADE and BALO}
\usage{
validate_bandcode_species(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with band code/species issues
}
\description{
Validate band code-species combinations. Make sure 4 and 8
are only used for species codes BADE and BALO
}
