% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bandsize}
\alias{validate_bandsize}
\title{Validate band size. Make sure there are no blanks. Make sure the only
values used are 0A, 0, 1, 1B, 1A, 1C, 2, 3, 3A, 3B}
\usage{
validate_bandsize(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with band size issues
}
\description{
Validate band size. Make sure there are no blanks. Make sure the only
values used are 0A, 0, 1, 1B, 1A, 1C, 2, 3, 3A, 3B
}
