% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bandsize_disp}
\alias{validate_bandsize_disp}
\title{Validate band size-disp combinations}
\usage{
validate_bandsize_disp(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with band size/disp issues
}
\description{
Validate band size-disp combinations
}
