% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bmlt}
\alias{validate_bmlt}
\title{Validate body molt. Allowable values: 0-4, blank}
\usage{
validate_bmlt(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with body molt issues
}
\description{
Validate body molt. Allowable values: 0-4, blank
}
