% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bp}
\alias{validate_bp}
\title{Validate BP (0-5, blank okay)}
\usage{
validate_bp(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with BP issues
}
\description{
Validate BP (0-5, blank okay)
}
