% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_bp_hs}
\alias{validate_bp_hs}
\title{Validate how sexed and BP for females. If sexed by BP, BP value cannot
be blank or 0}
\usage{
validate_bp_hs(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with BP/how sexed issues for females
}
\description{
Validate how sexed and BP for females. If sexed by BP, BP value cannot
be blank or 0
}
