% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_captime}
\alias{validate_captime}
\title{Validate cap time.Allowed values include: 650 to 1300. Flag all
other values. Other values may happen only if there is a note, sometimes
songbirds are caught during owls, hawk trapping, etc. All values should end
in 0's}
\usage{
validate_captime(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with cap time issues
}
\description{
Validate cap time.Allowed values include: 650 to 1300. Flag all
other values. Other values may happen only if there is a note, sometimes
songbirds are caught during owls, hawk trapping, etc. All values should end
in 0's
}
