% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_cp}
\alias{validate_cp}
\title{Validate CP (0-3 allowed, blank okay)}
\usage{
validate_cp(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with CP issues
}
\description{
Validate CP (0-3 allowed, blank okay)
}
