% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_cp_hs}
\alias{validate_cp_hs}
\title{Validate how sexed and CP for males. If sexed by CL, CP value cannot
be blank, 0, or 1 (i.e. CP must = 2 or 3)}
\usage{
validate_cp_hs(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with CP/how sexed issues for males
}
\description{
Validate how sexed and CP for males. If sexed by CL, CP value cannot
be blank, 0, or 1 (i.e. CP must = 2 or 3)
}
