% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_day}
\alias{validate_day}
\title{Validate day. Valid: 1-31. no blanks except for BADE/BALO}
\usage{
validate_day(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with day issues
}
\description{
Validate day. Valid: 1-31. no blanks except for BADE/BALO
}
