% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_day_species}
\alias{validate_day_species}
\title{Validate day-species combinations}
\usage{
validate_day_species(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with day/species issues
}
\description{
Validate day-species combinations
}
