% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_disp}
\alias{validate_disp}
\title{Validate disp. Allowable values include: M,O,I,S,E,D,T,W,B,L,P, blank}
\usage{
validate_disp(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with disp issues
}
\description{
Validate disp. Allowable values include: M,O,I,S,E,D,T,W,B,L,P, blank
}
