% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_disp_status}
\alias{validate_disp_status}
\title{Validate disp-status combinations. Any bird
with a letter in disp should have a note explaining why and the status should
say 500}
\usage{
validate_disp_status(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with disp/status issues
}
\description{
Validate disp-status combinations. Any bird
with a letter in disp should have a note explaining why and the status should
say 500
}
