% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ey}
\alias{validate_ey}
\title{Validate EY in how aged. EY in the How Aged columns should only be used for species codes SPTO, DOWO,
NOFL, RSFL, HAWO, DEJU, ORJU, SCJU, UDEJ --flag any other species that use
this with note, Check in Pyle to confirm that this species can be aged by eye
color}
\usage{
validate_ey(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with EY issues
}
\description{
Validate EY in how aged. EY in the How Aged columns should only be used for species codes SPTO, DOWO,
NOFL, RSFL, HAWO, DEJU, ORJU, SCJU, UDEJ --flag any other species that use
this with note, Check in Pyle to confirm that this species can be aged by eye
color
}
