% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_fat}
\alias{validate_fat}
\title{Validate fat 0-5, blank are allowed. 6 fat is okay but only if
there's a note}
\usage{
validate_fat(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with fat issues
}
\description{
Validate fat 0-5, blank are allowed. 6 fat is okay but only if
there's a note
}
