% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ffmolt}
\alias{validate_ffmolt}
\title{Validate flight feather molt. Allowable values: N, S, J, A, blank}
\usage{
validate_ffmolt(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with ffmolt issues
}
\description{
Validate flight feather molt. Allowable values: N, S, J, A, blank
}
