% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ffwear}
\alias{validate_ffwear}
\title{Validate flight feather wear. Allowable values: 0-5, blank}
\usage{
validate_ffwear(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with ffwear issues
}
\description{
Validate flight feather wear. Allowable values: 0-5, blank
}
