% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ha_ffmlt}
\alias{validate_ha_ffmlt}
\title{Validate how aged-ffmolt combinations.
If "how aged" says MR, FF molt must be S or J (can't be blank, N, or A)}
\usage{
validate_ha_ffmlt(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with how aged/ffmolt issues
}
\description{
Validate how aged-ffmolt combinations.
If "how aged" says MR, FF molt must be S or J (can't be blank, N, or A)
}
