% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ha_ffwear}
\alias{validate_ha_ffwear}
\title{Validate how aged-ffwear combinations.
If "how aged" says FF then FF Wear cannot be blank}
\usage{
validate_ha_ffwear(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with how aged/ffwear issues
}
\description{
Validate how aged-ffwear combinations.
If "how aged" says FF then FF Wear cannot be blank
}
