% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ha_ha2}
\alias{validate_ha_ha2}
\title{Validate how aged-how aged 2 combinations}
\usage{
validate_ha_ha2(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with ha/ha2 issues
}
\description{
Validate how aged-how aged 2 combinations
}
