% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_ha_skull}
\alias{validate_ha_skull}
\title{Validate how aged and skull combinations}
\usage{
validate_ha_skull(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with how aged/skull issues
}
\description{
Validate how aged and skull combinations
}
