% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_hs_hs2}
\alias{validate_hs_hs2}
\title{Validate how sexed-how sexed 2 combinations}
\usage{
validate_hs_hs2(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with hs/hs2 issues
}
\description{
Validate how sexed-how sexed 2 combinations
}
