% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_location}
\alias{validate_location}
\title{Validate location. Make sure there are no blanks}
\usage{
validate_location(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with location issues
}
\description{
Validate location. Make sure there are no blanks
}
