% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_month}
\alias{validate_month}
\title{Validate month. Valid: 2-11. No blanks except for BADE BALO}
\usage{
validate_month(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with month issues
}
\description{
Validate month. Valid: 2-11. No blanks except for BADE BALO
}
