% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_month_species}
\alias{validate_month_species}
\title{Validate month-species combinations}
\usage{
validate_month_species(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with month/species issues
}
\description{
Validate month-species combinations
}
