% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_muscle}
\alias{validate_muscle}
\title{Validate muscle. 2.5,3,4,5, blank allowed. 1 or 2 are allowed but
MUST have a note, otherwise it's likely a type-o (check hard copy)}
\usage{
validate_muscle(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with muscle issues
}
\description{
Validate muscle. 2.5,3,4,5, blank allowed. 1 or 2 are allowed but
MUST have a note, otherwise it's likely a type-o (check hard copy)
}
