% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_net}
\alias{validate_net}
\title{Validate net. Allowable values: 1-12, blank. Some exceptions allowed
with a note, e.g. owl nets but we should flag those exceptions anyway to
make sure someone checks them}
\usage{
validate_net(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with net issues
}
\description{
Validate net. Allowable values: 1-12, blank. Some exceptions allowed
with a note, e.g. owl nets but we should flag those exceptions anyway to
make sure someone checks them
}
