% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_notes}
\alias{validate_notes}
\title{Validate notes. Check that notes that mention either, FF, flat flies, or mites, lice,
louse, mite, fly have a Y for parasite column}
\usage{
validate_notes(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with notes issues
}
\description{
Validate notes. Check that notes that mention either, FF, flat flies, or mites, lice,
louse, mite, fly have a Y for parasite column
}
