% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_parasites}
\alias{validate_parasites}
\title{Validate parasites. If there is a Y in the parasites column there needs to be a note}
\usage{
validate_parasites(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with parasite column issues
}
\description{
Validate parasites. If there is a Y in the parasites column there needs to be a note
}
