% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_sex}
\alias{validate_sex}
\title{Validate sex column. Acceptable values= M F U--flag all the blanks}
\usage{
validate_sex(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with sex issues
}
\description{
Validate sex column. Acceptable values= M F U--flag all the blanks
}
