% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_sex_hs}
\alias{validate_sex_hs}
\title{Validate how sexed and sex combinations.  Allowable values include: PL,
EY,FF,MB,PC,LP,NL,MR,SK,TS, (blank only in second field, or for age 0)}
\usage{
validate_sex_hs(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with hs/sex issues
}
\description{
F: PL,BP,WL--first HS field can NOT be blank
}
\details{
M: PL,CL,WL--first HS field can NOT be blank

U: always blank, or IC, If not blank, check hard copy for errors or
white-out. If sex is whited out, leave as U. Check fields above and below to
make sure there's not a data entry error
}
