% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_species}
\alias{validate_species}
\title{Validate species column. Refer to master species list to update}
\usage{
validate_species(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with species issues
}
\description{
Validate species column. Refer to master species list to update
}
