% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_status}
\alias{validate_status}
\title{Validate status. Allowable values for new bands: 300, 500. Blank is NOT valid}
\usage{
validate_status(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with status issues
}
\description{
Validate status. Allowable values for new bands: 300, 500. Blank is NOT valid
}
