% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_status_500}
\alias{validate_status_500}
\title{Validate status 500s. ALL status 500's MUST have text in the note column and a letter in the disp
column i.e. Note and Disp columns cannot be blank}
\usage{
validate_status_500(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with status 500 issues
}
\description{
Validate status 500s. ALL status 500's MUST have text in the note column and a letter in the disp
column i.e. Note and Disp columns cannot be blank
}
