% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_tail}
\alias{validate_tail}
\title{Validate tail. Check if tail is below 30 or above 200}
\usage{
validate_tail(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with tail issues
}
\description{
Validate tail. Check if tail is below 30 or above 200
}
