% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_weight}
\alias{validate_weight}
\title{Validate weight. Flag anything under 5 but GCKI or BCHU RUHU CAHU okay or over 200 raptors would
be a rare exception}
\usage{
validate_weight(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with weight issues
}
\description{
Validate weight. Flag anything under 5 but GCKI or BCHU RUHU CAHU okay or over 200 raptors would
be a rare exception
}
