% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_wing}
\alias{validate_wing}
\title{Validate wing. Check if wing is below 30 or above 200}
\usage{
validate_wing(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with wing issues
}
\description{
Validate wing. Check if wing is below 30 or above 200
}
