% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_year}
\alias{validate_year}
\title{Validate year. No blanks. Allowable values are any valid year between
1997 and current year except BADE BALO}
\usage{
validate_year(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with year issues
}
\description{
Validate year. No blanks. Allowable values are any valid year between
1997 and current year except BADE BALO
}
