% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_year_species}
\alias{validate_year_species}
\title{Validate year-species combinations}
\usage{
validate_year_species(df)
}
\arguments{
\item{df}{bird data frame}
}
\value{
data frame of rows with year/species issues
}
\description{
Validate year-species combinations
}
